Make CoordinateEncoder an abstract class
