+++ server/src/main/java/org/elasticsearch/lucene/spatial/CoordinateEncoder.java
@@ -8,30 +8,134 @@
 
 package org.elasticsearch.lucene.spatial;
 
+import org.apache.lucene.geo.GeoEncodingUtils;
+import org.apache.lucene.geo.XYEncodingUtils;
+import org.elasticsearch.common.geo.GeoUtils;
+
 /**
- * Interface for classes that help encode double-valued spatial coordinates x/y to
+ * Abstract API for classes that help encode double-valued spatial coordinates x/y to
  * their integer-encoded serialized form and decode them back
  */
-public interface CoordinateEncoder {
+public abstract class CoordinateEncoder {
 
-    CoordinateEncoder GEO = new GeoShapeCoordinateEncoder();
-    CoordinateEncoder CARTESIAN = new CartesianShapeCoordinateEncoder();
+    /** Encodes lat/lon values into / from lucene encoded format */
+    public static final CoordinateEncoder GEO = new GeoShapeCoordinateEncoder();
+    /** Encodes arbitrary x/y values in the float space into / from sortable integers */
+    public static final CoordinateEncoder CARTESIAN = new CartesianShapeCoordinateEncoder();
 
     /** encode X value */
-    int encodeX(double x);
+    public abstract int encodeX(double x);
 
     /** encode Y value */
-    int encodeY(double y);
+    public abstract int encodeY(double y);
 
     /** decode X value */
-    double decodeX(int x);
+    public abstract double decodeX(int x);
 
     /** decode Y value */
-    double decodeY(int y);
+    public abstract double decodeY(int y);
 
     /** normalize X value */
-    double normalizeX(double x);
+    public abstract double normalizeX(double x);
 
     /** normalize Y value */
-    double normalizeY(double y);
+    public abstract double normalizeY(double y);
+
+    private static class GeoShapeCoordinateEncoder extends CoordinateEncoder {
+
+        @Override
+        public int encodeX(double x) {
+            if (x == Double.NEGATIVE_INFINITY) {
+                return Integer.MIN_VALUE;
+            }
+            if (x == Double.POSITIVE_INFINITY) {
+                return Integer.MAX_VALUE;
+            }
+            return GeoEncodingUtils.encodeLongitude(x);
+        }
+
+        @Override
+        public int encodeY(double y) {
+            if (y == Double.NEGATIVE_INFINITY) {
+                return Integer.MIN_VALUE;
+            }
+            if (y == Double.POSITIVE_INFINITY) {
+                return Integer.MAX_VALUE;
+            }
+            return GeoEncodingUtils.encodeLatitude(y);
+        }
+
+        @Override
+        public double decodeX(int x) {
+            return GeoEncodingUtils.decodeLongitude(x);
+        }
+
+        @Override
+        public double decodeY(int y) {
+            return GeoEncodingUtils.decodeLatitude(y);
+        }
+
+        @Override
+        public double normalizeX(double x) {
+            return GeoUtils.normalizeLon(x);
+        }
+
+        @Override
+        public double normalizeY(double y) {
+            return GeoUtils.normalizeLat(y);
+        }
+    }
+
+    private static class CartesianShapeCoordinateEncoder extends CoordinateEncoder {
+
+        private int encode(double value) {
+            if (value == Double.NEGATIVE_INFINITY) {
+                return Integer.MIN_VALUE;
+            }
+            if (value == Double.POSITIVE_INFINITY) {
+                return Integer.MAX_VALUE;
+            }
+            return XYEncodingUtils.encode((float) value);
+        }
+
+        private double decode(int value) {
+            if (value == Integer.MIN_VALUE) {
+                return Double.NEGATIVE_INFINITY;
+            }
+            if (value == Integer.MAX_VALUE) {
+                return Double.POSITIVE_INFINITY;
+            }
+            return XYEncodingUtils.decode(value);
+        }
+
+        @Override
+        public int encodeX(double x) {
+            return encode(x);
+        }
+
+        @Override
+        public int encodeY(double y) {
+            return encode(y);
+        }
+
+        @Override
+        public double decodeX(int x) {
+            return decode(x);
+        }
+
+        @Override
+        public double decodeY(int y) {
+            return decode(y);
+        }
+
+        @Override
+        public double normalizeX(double x) {
+            return x;
+        }
+
+        @Override
+        public double normalizeY(double y) {
+            return y;
+        }
+    }
 }
